Refactor `fetchAndRenderPath` to append result to DOM
This commit refactors fetchAndRenderPath so that it is the repsonsibility
of the caller to supply a targetParentElement that the newly rendered tree
will get appended to.

PathDepth is now calculated based on the data attribute of the parent.

updateView is no longer responsible for identifying the preexisting
path segment already rendered in the dom, rather, it calls fetchAndRenderPath
no matter what, and supplies a target parent dom node, and it is
fetchAndRenderPath's job to identify if the first path element already exists,
and if so, recursively call itself with the remaining path segment and preexisting
node as parent node. This was preferable to having fetchAndRenderPath do this
check defensively and also have updateView do it.

A new helper `sectionElementOfRelativePath` is introduced, and the existing
helper `sectionElementOfPath` is rewritten to use it.