Extract setPath and storeLinkSelectionInSession into new function
`displayPath` calls `setPath` and `storeLinkSelectionInSession`, and the call to `setPath` must precede the call to `storeLinkSelectionInSession` because the data is stored in the session based on what the current URL is, and so if session storage occurs before the URL is updated, it will store the link selection for the wrong paragraph causing corruption when that link isn't found in that paragraph. This is solved by extracting the two operations into a new function with a comment that makes it clear that the order is significant.