Add section placeholders during render to avoid duplicate child elements
Previously, it was possible for a paragraph that had multiple references to the same global child to render and append multiple versions of it because you could have multiple calls to fetchAndRenderPath before any of them appended an element to the parent node, thus neither subroutine knew about the other and both would append their duplicate child. This fix causes a placeholder element to be appended before the asynchronous network call which is then decorated with the data from the server, but in the mean time serves as a signal to other global links in that paragraph not to render and append the same child paragraph.