Eager render global children upon navigating to subtopic
Previously, eager render of global children was caused at render time when rendering a global link that was in the terminal paragraph of the current path. However, if the user navigated to a new terminal paragraph, global children were not eagerly rendered.

Now, navigating to a new paragraph after page-load will trigger eager render of global children in that paragraph.