Simplify renderDomTree code by separating types of global child render
Previously, there was one code path that handled rendering the children of open global links, and eager render of global children of the terminal paragraph of the path. However, handling both of these cases simultaneously required a lot of abstraction in the code, because they weren't very similar. Now, the cases are handled by two much simpler code paths.