Power zoom command with new lazy link instantiation option
Previously, there was no good way to instantiate a link object if the link needed wasn't already in the DOM. This was relevant for example when the zoom command would try to select the same link at a new path. We needed a way for someone to instantiate a link object that would only later refer to an extant DOM node.

In this commit, we create three paths for instantiating a link object, one can provide a DOM element, a metadata object, or a selector callback. Providing one of the three should not trigger generation of any of the other three until they are actually requested, so metadata should not be generated for a linkElement until #metadata is called, and a linkElement should not be looked up for metadata or a selector callback until #element is called.

This commit adds a `transferDataset` method to Paragraph and Link in order to improve visibility in the developer console. A #element setter is added to do this automatically. Getters for the element's dataset are replaced by accessing the new instance variables directly.

This commit adds a root getter for Paragraph.

This commit adds a getter on Link to retrieve relative link number in the enclosing paragraph.

The logic for taking metadata from a DOM element is extracted into the #collectMetadata class method.

Paragraph#containingLink will now error if the argument is not of type link.

Link#element used to check for link#linkElement and then link#metadata, but the call to #metadata could end up looking for a #linkElement in order to get the metadata, so to simplify things Link#element checks for #metadataObject which just checks if the user passed in a metadata argument, preventing potential cycles.

Caching is added for Link#relativeLinkNumber and Paragraph#links

Temp link bug