Add parent link child preview behavior
Previously, selecting a parent link did not display the child paragraph until the user "opened" the parent link. Now, selecting a parent link and opening it are the same thing.

updateView already takes a path and link argument, now we check the path of the link when selected, and if it contradicts the provided path, we go with the link. If so, it would seem as if the caller should only have to provide the link if from that we can derive the path, and there is only a chance of contradictory inputs, but the path must be provided in the case of "selectALink" wherein some link needs to be selected, but we are not sure which one, and depending on the link's type, selecting it might change the given path, which can't be known until after rendering, which requires having requested the data, which requires having had a path to request in the first place. Additionally, we need a path provided in the initial state where no link is selected.

Helpers were added to achieve this goal:

`Link#matches(otherLink)` is a boolean function that checks whether the basic metadata of two links match, regardless of whether those links occur at the same path. This is used in the "zoom" keyboard shortcut.

 `Link#atNewPath(newPath)` takes a path and returns the link that matches the given link at that new path. This also is used in the zoom shortcut.

`Link#targetPath` gives the path of the target paragraph of a local or global link. For eg a URL link it returns null.

`Link#pathWhenSelected` is similar to #targetPath, except even for URL links it returns a value, which is the path of the paragraph in which the URL link appears. `pathWhenSelected` is intended to equal the path of the lowest visible paragraph on the page when the given path is being viewed, which for a URL link is the enclosing paragraph, but for a parent link is now the previewed child paragraph.

`Link#hasTarget(topic[, subtopic])` returns a boolean function that checks whether a given link argument has a given target topic and subtopic, or if subtopic is not given, whether it's target is the topic subtopic of the given topic tree.

`Link#contradicts(path)` checks whether the path given matches the path that would be displayed if the given link was selected. Contradiction is possible eg in a case where we wish to "selectALink" ie any link, and the link selected turns out to be a parent link, retroactively changing the displayed path once selected. Contradiction in such a case is desired functionality.

`Paragraph#topicParagraph` returns the paragraph that is the root of the topic tree of the given paragraph.

`Path#parentHasConnectingLink(parentElement, pathToDisplay)` checks whether a given parentElement contains a link such that it could validly have the given path rendered below it. This addresses a bug in which paths which had no connecting link path could nevertheless be rendered so long as the URL matched the incorrect DOM state.

The css of an open and selected link used to combine the bold of the open links and the highlighting of the selected links, but now that all selected links are open, the highlighting alone is chosen for the combined state.

`Path#parentParagraph` gives the parent paragraph of the given path's paragraph.

`Paragraph#root` is renamed `Paragraph#pageRoot` to clarify that this method returns the root paragraph of the entire displayed page, not just a particular subtree or name of something.

`Link#selectALink` now takes a path argument because sometimes the callback was triggering before the new path was set, causing the link selection to occur on the previous DOM state, contradicting the new path eg in the selection of a global link causing full-page redirect.

`Link#localPathWhenSelected` gives the portion of the path that concerns the selected link, if local or URL link, that is just the last path segment, but if a global link is being previewed, that includes the last two segments. This is used in the zoom shortcut.

temp

temp