Remove concept of eager render of global children
After doing profiling it turns out that the render time of global child paragraphs when selected is negligible so long as eager loading has been done for the network request upon encountering the global link during render.

We no longer need to use placeholder elements because there is no concern of multiple global links in the same paragraph causing multiple copies of the same global child to be appended because that was only a concern when every global child was being appended concurrently, and different renderings for the different redundant global links might not know about one another, however now, all rendering is initiated by user selection, and so one of the duplicate global links must be selected before the other, and will fully render before the UI permits a second selection, and so the existing `preexistingElement` code in fetchAndRenderPath will suffice to ensure further paths that begin with the given global child reuse the existing section element during render.