Remove redundant parents from client
Previously, a two local references within a topic tree would both become links, and the second would redirect to the first. Now, this behavior is being deprecated in favor of using import references within the tree, which will be permitted in a future update. Redundant parent links will also be removed from the parser in a future commit and will become a no-op or error.