Support hash Urls
Following the update of the CLI, projects can now be built with a "hashUrls" option, which this commit implements on the front-end. When `hashUrls` is enabled, Canopy paths are prefixed with '/#/'. This is helpful if one is trying to run a Canopy website on a static assets server that does not support symlinks, causing errors when a user tries to initially navigate to eg "example.com/TopicA/TopicB", and there is no such directory.

This commit also fixes the previously malfunctioning "projectPathPrefix" option, which allows one to serve the Canopy website on a subdirectory of a website eg "example.com/subdirectory/TopicA".

Both options can be used in conjunction, eg "example.com/subdirectory/#/TopicA