Use `link.pathToDisplay` to simplify key handlers
Most of the time, the paragraph path that should be displayed is a simple function of the new link that is selected. Each of the link types has a certain behavior for what path should be displayed when they are selected. In theory, `updateView` should therefore be able to take just a link argument. However, in practice, there are two edge cases that prevent this, the case where one presses up on a link in the root paragraph thus unselecting all links, and the case where one is requesting a new path that isn't in the DOM yet, and so there is no way to get a hold of a link DOM element in order to specify which one should be selected, and so one instantiates a link with a callback to find the link on page load. In these cases either there is no link to derive a path from, or, we do not know before page load what the link will be, and therefore we don't know what the path will be at initial page load, whereas we need a path before render to request data from the server.

So, to simplify things, key handlers in general should not derive the link to select and the path to display separately, which introduces room for inconsistencies, but rather should select the link, and then derive the path from `link.pathToDisplay` ie the path to display when that link is selected. This resolves eg a bug where selecting a global link without children caused  the current link to remain selected, but the target topic to be appended to the current path as if advancing to a new child link. If paths are derived from links then this class of error cannot occur.

This commit also refactors deep nesting in `moveDownOrRedirect`

This commit removes the path contradiction notice as this is now the standard behavior in situations where updateView must be called before the final link selection is known, after which the link selection changes the current path.

`updateView` now returns a promise so that callers can handle errors, eg if selectALink is used in a situation where the resulting page has no links, eg return on a global link with no children, which should not fail but rather redirect without selecting any link.