Rewrite bulk mode to support many files per path
Previously, every file needed to have its own path in the canopy bulk file.

```
A/B/C:

File

A/B/C:

Other File

```

This commit introduces a change where in 1) multiple files can be declared at each path, and 2) files and paths are distinguished with special characters. So, the above example would now look like this:

```
[A/B/C]

* File

* Other file

```

The benefit of this is that one doesn't have to write the full path, and as in the previous version, filenames are inferred from the first paragraph's key.

Files that lack keys are added to the path's "category file", ie for `A/B/C`, they would be added to `A/B/C/C.expl`.