Handle invalid connecting link gracefully
Previously, if rendering the path A/B/C, and paragraph B didn't have a link to C, we would throw an error and stop. This throws away all the progress made rendering the path A/B, which had no problems until C. In this commit, `fetchAndRenderPath` detects the problem with A/B/C, but doesn't throw away the progress made in rendering A/B, and aborts the rendering process and allows the display code to notice the missing C and redirect to A/B using the `tryPathPrefix` function called in `displayPath`.