Change link inlining behavior
Previously, clicking or pressing return on a selected global link would redirect you to that topic as the page root. It seems however that people are more likely to expect inlining behavior to be the norm unless they opt out of it explicitly.

Therefore, now, all click and return key handlers will assume inlining unless the user is pressing the alt/option button, which will now indicate that they wish to either redirect to a given global / import reference path, or in the case of a local link, zoom to its path segment. Holding command will do the same but will open a new window.

Clicking a link that isn't open now also selects that link.

This commit also introduces a change wherein if paragraph A has an import reference to paragraph A/B#C, when A/B#C is open, the import reference in A is opened, but if any other paragraph in A/B is selected it is not. This is to prevent the counter-intuitive UI experience where a user clicks the import reference, and the import reference doesn't open, but rather the adjacent topic link opens. Now, both will open.

This commit also fixes href generation on link elements which was previously incorrect. Hrefs on links reflect the literal target of the link, rather than the contextual path of the paragraph's position in the current session. This is because if JS is enabled it is irrelevant what the href is because we are capturing the click event, in which case the href might as well just be something succinct and clear.

"Parent Link" is replaced with "Local link" in click handlers.

Link#atNewPath previously was synchronous and required the new link to already exist, which required wrapping it in a callback and instantiating a link with the callback. Now, Link#atNewPath does that for you, and returns a link instantiated from a callback that is expecting to be invoked only after page render.

This commit also changes the behavior of a page that loads with a path already open. Previously, we still waited for the first keystroke to select a link, even though many links were open. Now, the lowest open link automatically becomes selected if no other link is selected. If no link is open, no link is selected.

This commit fixes a bug where if projectPrefix was null, the leading slash of the current path was omitted, leading to compounding paths on navigation.

This is a table of the current behaviors:

+-----------------------+----------+----------------+-------------------------------+
|                       | Global   | Local          | Import                        |
+-----------------------+----------+----------------+-------------------------------+
| Down                  | Inline   | Inline         | Inline, select target         |
+-----------------------+----------+----------------+-------------------------------+
| Return                | Redirect | Inline         | Redirect, select target       |
+-----------------------+----------+----------------+-------------------------------+
| Click                 | Inline   | Inline         | Select import, display target |
+-----------------------+----------+----------------+-------------------------------+
| Option-Down           | N/A      | N/A            | N/A                           |
+-----------------------+----------+----------------+-------------------------------+
| Option-Return         | Redirect | Zoom on child  | Redirect to target            |
+-----------------------+----------+----------------+-------------------------------+
| Option-Click          | Redirect | Zoom on child  | Redirect to Target            |
+-----------------------+----------+----------------+-------------------------------+
| Command-down          | N/A      | N/A            | N/A                           |
+-----------------------+----------+----------------+-------------------------------+
| Command-return        | Redirect | Inline         | Inline, select target         |
+-----------------------+----------+----------------+-------------------------------+
| Command-click         | Inline   | Inline         | Inline, select target         |
+-----------------------+----------+----------------+-------------------------------+
| Command-option-Down   | N/A      | N/A            | N/A                           |
+-----------------------+----------+----------------+-------------------------------+
| Command-option-return | Redirect | Zoom           | Redirect to target            |
+-----------------------+----------+----------------+-------------------------------+
| Command-option-click  | Redirect | Zoom           | Redirect to target            |
+-----------------------+----------+----------------+-------------------------------+