Allow question mark keys
Previously, in an `expl` file, the definition of a key was a string at the beginning of a paragraph ending with `:` and followed by further text. Now, question mark is a valid alternative to demarcate the divider between key and paragraph.

This commit also removes the `extractKeyAndParagraph` helper and replaces it with a Paragraph class.