Refactor parser code and move into CLI
This commit introduces a number of changes to the Canopy JS parser.

First, the parser is moved from webpack to node because it is never served over a browser so building a single asset is unnecessary.

Second, the parser moves from natural language references to media wiki style references eg [[link]]. This is due first to performance concerns, as natural language references take O(NS^2) to parse, where N is the total size of the project and S is the size of the given sentence. Secondly, explicit references allow us to sometimes use a topic name without intending to linkify it, and also makes it easer for someone reading the `.expl` file to know what is intended as a link without knowing the sum total of all existing topics and subtopics.

Import references are separated from global references as an independent type of token, and can be defined implicitly without specifying topic if there is exactly one matching topic in the same parsed text. Selecting an import reference now displays the target path below it, and when having selected an import reference, the path displayed should be to the target paragraph, but the URL should be to the import reference's enclosing topic, so that if the user proceedes to select the import reference's target, there is a URL change, such that if the user then does a browser back, the import reference should be selected.

The parser no longer needs to make 3 passes because we no longer need to know whether an import reference is valid at the time of tokenizing, rather, we build an object of `ImportReferencesToCheck`, and keep track of `subtopicParents` ie which subtopics are subsumed by their parent topic, and then at the end, validate that every import reference is to a subtopic that was subsumed.

Rather than breaking the text into sentences, now we parse the entire text at once, solving the issue of false positives eg "Dr. " as a sentence ender.

Link matching is now case-insensitive, and topic and subtopic names are all-caps in `topicSubtopics` for this reason.

The parser is no longer invoked from the command line but rather exports a function `buildProject` that is invoked in the CLI's build command.

Tests are updated including those relevant to previous commits.

Parser helpers are consolidated into one file.