Simplify and enhance style support
Canopy does not fully support markdown styles because each token is parsed individually, and so you can't have eg an italics block that spans multiple text or link tokens. However, the characters ` _ * and ~ do work within the text of a single text or link token.

Previously the implementation was more complicated and entailed identifying chunks of text that had the same combination of style attributes and rendering each chunk separately with its own style tags, even if the style tags from one to another were continuous, eg _a * b * c_ would render to <i>a </i><i><b> b </b></i><i> c</i> rather than the simpler <i> a <b> b </b> c </i>. This commit simplifies the logic and produces the more expected DOM output.

Furthermore, this commit handles the case where something that looks like a style token must be rejected as such because it lacks a later closing style token.

This commit also removes the term "markdown" from the codebase because the Canopy implementation of `_` and `*` characters goes away from the markdown standard in that `_` is always italics and `*` is always bold.