Extract topic names into class
This commit introduces a `TopicName` class to the parser code. The Topic class keeps track of the different stylistic permutations of the topic and subtopic strings the user enters in the `expl` files.

Whereas previously the code would pass around strings, now it passes around Topic objects and the caller invokes eg `.caps` if the caller wants caps, etc.

Remove markdown support for topic names

Previously, topics could have italicized or underlined text by using markdown characters like _underscores_. This led to code complexity and conceptual complexity, such as which link texts should match a topic that has markdown, etc.

This commit removes the concept of markdown characters from topic names. In theory headers and link texts can contain markdown characters, but if they do, these are just regular characters as far as matching is concerned. Whether the DOM will render the characters as style tags is now unspecified behavior, although for the moment it will continue to work.

The concept of a topic name for the purposes of matching is now an all caps string of characters. Markdown characters are not special characters. The behavior of quotation marks in topic names requires further investigation.