Remove null link behavior
Previously if a link was instantiated with a link selector callback
and later the callback didn't return any value, we would continue
silently with a "null link", and methods that expected a link element
had to check for the presence of a link. It seemed like this would
be hard to remember. Alternatively, we could have the callers to
updateView in situations where a null returning link selector
callback seemed likely to catch manually and recall updateView
without a link selection, but this also required callers to remember
this edge case.

So, this commit introduces a third option, wherein updateView always
knows to recall displayPath without the link argument if it gets
precisely this error, which allows callers to updateView to be
none the wiser, and allows writers of methods on Link to assume the
presence of link.element.