Bug: Upon matching token, do not add current char to buffer
If we don't find a token match at characters[i], we want to add
characters[i] to the buffer and continue. The buffer will become a
text token when we do find a token, or reach the end of the text.

If however we do find a token, we don't want to add a character
to the buffer because that next character might be part of a new
token, so we advance `i` to the last character of the previous token
and continue the loop so that on the next iteration, we begin with `i`
pointing to the first character after the token, and begin to check
for token matches.