Add bottom padding to last paragraph
Previously, once the screen was full, paragraphs would be added
to a full screen, making it unclear that downward motion was
occuring. Now, padding is added to the last element. That way,
being at the bottom of the page looks different from having
scrolled up and looking at the middle. Hopefully, this makes it
clearer that the paragraphs being added are being added to the
bottom, and that the page is moving relative to the viewer,
rather than the viewer moving relative to the paragraphs (ie
scrolling.)

This commit introduces a new Paragraph#select method and calls it in the
displayPath function. This indicates which is the lowest paragraph on
the page so that we can selectively add bottom padding to it
specifically.