Remove quotation marks from JSON filenames
Previously, quotation marks could be used in topic names and they would
get added to the filename. This broke file serving in Express, so now
file names have the quotation marks and other problematic characters
removed.

This file also adds errors if topics are added that will have the same
filename, or subtopics are added that will have the same URL. (Topics
have a stricter test for uniqueness, because they cannot differ on
quotation marks etc as these are stripped from filenames, but subtopics
which do not need filenames can differ on quotation marks.)