Rename promise array to displayBlockingPromises
The render code passes around an array of promises and any action
that is blocking the display of the page gets added to it. Hopefully
this name is a clearer indication that the array is a list of tasks
that must be completed before moving on to the display process.