Build test project as part of browser tests
Previously, when running Playwright locally it would rely on the
test project already being built, which the developer might only have
done several commits ago. This new commit makes building the project
be part of the browser tests so that it will capture any errors in the
build as well.