Import reference should beat redundant local reference
Previously, anything that could be a subtopic of the current topic
was considered to be a local reference, even if that meant throwing
an error for a redundant local reference.

Now, if a certain link would be a redundant local reference, but it
could also be an import reference, we assume it is an import reference.

This allows users to use abbreviated import references even a subtopic
with the same name as the target subtopic also belongs to the enclosing
topic.

This however risks interpreting as an import reference something the
user actually intended as a redundant local reference, where they should
actually receive an error. Hopefully they will recognize that the
behavior of the import reference is unexpected. Hopefully also it is
unlikely that a user who is confused about redundant local references
hasn't yet received an error about it and this is the first time they
need one.

(~Perhaps by the same logic, if we come across a local reference A in a
certain subtopic, and A could have also been an import reference, but
the interpretation as local reference had priority, and then we come
across a redundant local reference B with the same text, but B cannot
also be an import reference, then retroactively A was a redundant local
reference, and so if it could have also been an import reference,
perhaps that interpretation should have won over raising an error. This
could be solved by keeping eg a localReferenceLedger that for every
local reference, checks if it could be an import reference, and if so,
stores information about what import reference it would be, so that if
it later ends up being a redundant local reference, we can use the
ledger to find the original local reference token in memory and swap
it out with an import reference. This however seems like a lot of
overhead for a case where the user has the option to either use
expanded import reference syntax for A, or just move the paragraph of A
down so that it is seen second.~) – (This functionality was introduced
in the following two commits.)

This commit also renames topicReferences to topicReferencesInText, and
adds a few explanatory comments to the matchers code.