Break implicit import reference ties with proximity
Canopy supports implicit syntax for import references, eg:

"I want to go to [[London]], [[England]]."

Here, even though the subtopic "Paris" has not specified its topic,
we infer from the presence of "France" that this is its topic.

However, what if we have two candidate topics? Eg.

"I would like to visit [[Columbus]], [[Ohio]] and [[London]], [[England]]"

Here, there is an England#London, and there is also an Ohio#London, so
which one is [[London]] a reference to?

Previously, this sentence would raise an error that [[London]] must be
rewritten as either [[England#London]] or [[Ohio#London]] explicitly.

This commit however introducing a defaulting behavior wherein we assume
an implicit import reference belongs to the closest global link that has
a subtopic of that name, proximity being defined as from the starting
index for global links before the given link, or as from the ending
index for global links. Eg

"I would like to visit [[Columbus]], [[Ohio]] and [[London]], [[England]]"

London is 5 characters before [[Ohio]], and 2 characters before
[[England]], so it should bind to England.

For cases where this heuristic is incorrect, the user can always
override the defaulting behavior by explicitly specifying the correct
topic eg [[Ohio#London]].