Retroactively convert local references to import if found redundant
Lets say we have a topic "Ohio", and we are defining Ohio#Ohio:

```
"I would like to go to [[London]], [[England]]"
```

Now, if Ohio has a subtopic `Ohio#London`, that first link will become a
local reference incorrectly.

Now, lets say that later we are defining `Ohio#Counties`:

```
"Ohio has a county called [[London]]"
```

Now, we have redundant local references. Previously, this would throw
an error. Now, the parser notices that while the second local reference
can't be an import reference, the prior local reference could have been,
and so it goes back and changes it to be an import reference, avoiding
a parser error due to redundant local reference.