Isolate parser from file system for easier testing
Previously, the parser code would read from the file system and would
write to the file system directly. This made it difficult to test.

Now, the parser code is given a Javascript object representing the
expl files it needs to read, and it returns an object of file paths and
file contents and an array of directory paths representing the required
writes that must be done to the file system. This lets us easily create
tests that describe a hypothetical set of expl files and assert on what
JSON files the parser should write in response, all without having to
deal directly with the file system.

This commit also adds tests over error messages the parser displays when
a project has invalid data.

This commit also adds a syntax for creating global references when a
given topic name coincides with a subtopic name. In such a case,
using [[Topic#Topic]] will force creation of a global reference.