Isolate bulk CLI from file system for easier testing
Previously `reconstructProjectFiles` took the bulk data file and made
changes to the files system in light of it. Now, the function has been
refactored into smaller components. `parseDataFile` takes the data file
and creates a provisional list of files to write. Then
`compareChangesWithFileSystem` looks at that list and looks at the file
system data and decides which writes are appends and which are
overwrites, and also which missing files are really deletions.

Because neither `parseDataFile` nor `compareChangesWithFileSystem`
do any reads or writes to the file system directly, it is easier to test
that they are working.