Create `--continue` and `--no-backup` canopy bulk options
Often in development, the canopy bulk file will cause an error when
parsed, and running bulk again will overwrite the dotfile. Now,
if something prevents the parsing of the file but it is still on disk,
one can run `canopy bulk --continue` and resume where one left off.

This commit also renames the option that disabled backup logging which
is now `-n` or `--no-backup`.