Consolidate eagerLoad and requestJson
Previously, we had very similar code for requesting JSON from the server
and for eager-loading it. Now, there is just one function requestJson,
whose behavior is to request the file and store the _promise_ in a
cache. That way, if the real request occurs after the eager load but
before the data returns, then the request is not made twice. If the
real request occurs after the data has returned, it will receive an
already-resolved promise and will be none the wiser.