Handle cases where `canopy bulk` search finds no results
Previously, there were several errors that occurred if a search returned
zero results. Grep returned an error code which crashed the CLI. Also,
the code previously didn't distinguish between an empty file list due to
the user specifying no query, which should trigger recall of the entire
project, versus an empty file list due to their having given a query
but one that returned no results, in which case they should get a blank
screen.

Additionally, previously if one tried to save a blank session it would
error due to not having a valid path, but now this is a valid state.

Additionally, now the blank option is incompatible with the search
options.

And previously the "Invalid path" error was uncaught when using the
pop-up editor mode, and now there is a try catch block.