Add tests for more click and keyboard scenarios
This commit introduces tests for scenarios like option-clicking a global
link, meta-clicking a global link, meta-option clicking a global link.

This commit introduces a playwright browser test for DFS functionality,
iterating through the local references of a topic with the tab key.

Previously pressing shift tab would iterate through DFS backwards,
but this complicated the code and the same effect can be achieved with
browser history so this commit removes that shortcut.

This commit also adds browser test for shift-up shortcut
Shift-up should redirect to the parent link of the topic root paragraph.
If we are currently in the top stack frame, it should do nothing.