Filenames and topics can contain quotation marks
Previously we removed quotation marks from file names, but the
file system and browser requests allow them, and removing them makes
bulk mode work confusingly, because the [A/B/C] path is populated based
on the folder names, and either we remove quotation marks and so it
shows up wrong in bulk mode, or we don't and it looks funny that we
remove the quotations only from the filename and not the directory path.

This test adds a feature spec to ensure that both the build and client
code can handle apostrophes and double quotes.

This commit also adds URL decoding for when a topic with quotes is
opened in a new tab, because even though the quotation is shown in the
URL properly, programmatic access to the URL can contain query encoded
quote characters instead, so this commit decodes URL paths by default.