Represent empty paths with empty path array
Previously empty paths were represented with the array [[''],['']],
but now we are using the simpler [], and code that checks for the path
array must know that having zero segments is a valid option.

This fixes an error where pressing browser back and ending up on an
empty path created a path string of ',' in the URL, because it was
converting the [[''],['']] to a string.