Scroll smoothly during link selection
This commit introduces better automatic scrolling when selecting a
link.

When we remove DOM elements from the bottom of the screen such that the
bottom of the body becomes visible, the screen jaggedly adjusts to put
the new bottom of the body at the bottom of the screen, and this change
isn't animated. So, this commit adds a padding of 60% of the viewport
to the lowest situated paragraph so that under normal conditions the
true bottom of the body is never visible, so that when content is
removed, the bottom of the body still does not become visible, such
that there is no forced change to the scroll position.

Additionally, once the page is rendered, we now do a 'smooth' scroll
that puts the top of the paragraph containing the selected link 30%
of the viewport, leaving space below the selected link's paragraph for
previewed paragraphs, and below that for whitespace.

This commit also makes Paragraph#select a non-getter method to make it
clearer that the operation has side-effects.