Bug: Popping state to empty path caused error
On initial page load in canopy.js, we were using Path.initial which does
a check of Path.current.empty and returns Path.default if it is, but we
were doing no such check for the pop-state listener, so the pop-state
listener always used Path.current even if it was empty, causing an
error.

This commit introduces several guards against this scenario.

1. The pop state listener now calls Path.initial just like the regular
page load code path.

2. updateView now automatically replaces Path.current with Path.default
if the former is empty.

3. tryPrefix in the displayPath code will now redirect to the default
path after exhausting guesses from prefixes of the given path, which if
the path is empty would cause an automatic redirect.

This commit also renames link.pathToDisplay to be the more verbose
link.paragraphPathWhenSelected, and moves pathForLink to be a method
on link called link.pathForUrl, to highlight that we can't use the
given path argument because sometimes the link's path for URL is
different than the provided path argument.

This commit also adds a new test for when an empty path is entered
manually and then it is returned to using browser back, and adds to the
prior link selection browser history test, verifying that not only link
selection persists but also the lack of link selection for a given path.

This commit also renames Path.initial to Path.currentOrDefault because
it is used in the pop-state handler which isn't really properly called
an "initial" path, because its in the middle of a session.